refactor(metadata): extract shared SQL row/timestamp codec
The sqlite and postgres stores carried byte-identical inode-row decoding
(fileRowToFileWithNlink), FILETIME timestamp conversion, and folded
block-ref JSON decoding. Move them into a shared internal/sqlcodec package
behind a Row interface both backends' scanned rows already satisfy, so each
store keeps only its handle encoder and its backend-specific block-ref
aggregate SQL.
