+++ pkg/metadata/store/sqlite/files.go
@@ -12,6 +12,7 @@ import (
 	"github.com/marmos91/dittofs/pkg/block"
 	"github.com/marmos91/dittofs/pkg/metadata"
 	"github.com/marmos91/dittofs/pkg/metadata/acl"
+	"github.com/marmos91/dittofs/pkg/metadata/store/internal/sqlcodec"
 )
 
 // ============================================================================
@@ -56,7 +57,7 @@ func (s *SQLiteMetadataStore) GetFile(ctx context.Context, handle metadata.FileH
 	`
 
 	row := s.queryRow(ctx, query, id, shareName)
-	file, err := fileRowToFileWithNlinkAndBlocks(row, true)
+	file, err := sqlcodec.FileRowToFileWithNlinkAndBlocks(row, true)
 	if err != nil {
 		return nil, mapDBError(err, "GetFile", "")
 	}
@@ -303,10 +304,10 @@ func (s *SQLiteMetadataStore) ListChildren(ctx context.Context, dirHandle metada
 			UID:          uint32(uid),
 			GID:          uint32(gid),
 			Size:         uint64(size),
-			Atime:        nanosToTime(atime),
-			Mtime:        nanosToTime(mtime),
-			Ctime:        nanosToTime(ctime),
-			CreationTime: nanosToTime(creationTime),
+			Atime:        sqlcodec.FiletimeToTime(atime),
+			Mtime:        sqlcodec.FiletimeToTime(mtime),
+			Ctime:        sqlcodec.FiletimeToTime(ctime),
+			CreationTime: sqlcodec.FiletimeToTime(creationTime),
 			Hidden:       hidden,
 		}
 		if len(objectIDRaw) > 0 {
@@ -321,15 +322,15 @@ func (s *SQLiteMetadataStore) ListChildren(ctx context.Context, dirHandle metada
 
 		// Recycle-bin metadata (#190): carried on DirEntry.Attr so trash
 		// enumeration via listing reflects recycle state without a re-read.
-		// deleted_at is BIGINT unix-nanoseconds; decode via nanosToTime.
+		// deleted_at is BIGINT unix-nanoseconds; decode via sqlcodec.FiletimeToTime.
 		if deletedAt.Valid {
-			t := nanosToTime(deletedAt.Int64)
+			t := sqlcodec.FiletimeToTime(deletedAt.Int64)
 			attr.DeletedAt = &t
 		}
 		attr.OriginalPath = originalPath
 		attr.DeletedBy = deletedBy
 
-		// Refs #532 (PR #536 review): mirror fileRowToFileWithNlink. A
+		// Refs #532 (PR #536 review): mirror sqlcodec.FileRowToFileWithNlink. A
 		// malformed ACL row is treated as "no ACL" rather than failing the
 		// whole listing — same lenient behaviour the GetFile path has had
 		// since the column was introduced.
@@ -463,7 +464,7 @@ func (s *SQLiteMetadataStore) GetFileByPayloadID(ctx context.Context, payloadID 
 	`
 
 	row := s.queryRow(ctx, query, string(payloadID))
-	file, err := fileRowToFileWithNlink(row)
+	file, err := sqlcodec.FileRowToFileWithNlink(row)
 	if err != nil {
 		return nil, mapDBError(err, "GetFileByPayloadID", string(payloadID))
 	}
