+++ pkg/metadata/store/sqlite/transaction.go
@@ -14,6 +14,7 @@ import (
 	"github.com/marmos91/dittofs/pkg/metadata"
 	"github.com/marmos91/dittofs/pkg/metadata/acl"
 	"github.com/marmos91/dittofs/pkg/metadata/store/internal/quota"
+	"github.com/marmos91/dittofs/pkg/metadata/store/internal/sqlcodec"
 )
 
 // Transaction retry policy (#1769). Under write contention DittoFS must
@@ -212,7 +213,7 @@ func (tx *sqliteTransaction) GetFile(ctx context.Context, handle metadata.FileHa
 	`
 
 	row := tx.tx.QueryRow(ctx, query, id, shareName)
-	file, err := fileRowToFileWithNlinkAndBlocks(row, true)
+	file, err := sqlcodec.FileRowToFileWithNlinkAndBlocks(row, true)
 	if err != nil {
 		return nil, mapDBError(err, "GetFile", "")
 	}
@@ -330,12 +331,12 @@ func (tx *sqliteTransaction) PutFile(ctx context.Context, file *metadata.File) e
 
 	// deleted_at is a BIGINT Windows-FILETIME column (#190), nullable: NULL marks
 	// a live node, a value records the recycle instant losslessly (same encoding
-	// as the other file timestamps — must use timeToNanos, not UnixNano, so it
-	// decodes back correctly via nanosToTime). Pass *int64 so a nil DeletedAt
+	// as the other file timestamps — must use sqlcodec.TimeToFiletime, not UnixNano, so it
+	// decodes back correctly via sqlcodec.FiletimeToTime). Pass *int64 so a nil DeletedAt
 	// writes SQL NULL.
 	var deletedAtArg *int64
 	if file.DeletedAt != nil {
-		n := timeToNanos(*file.DeletedAt)
+		n := sqlcodec.TimeToFiletime(*file.DeletedAt)
 		deletedAtArg = &n
 	}
 
@@ -353,8 +354,8 @@ func (tx *sqliteTransaction) PutFile(ctx context.Context, file *metadata.File) e
 		// Row exists; update it in place.
 		if _, err := tx.tx.Exec(ctx, updateQuery,
 			file.Type, file.Mode, file.UID, file.GID, file.Size,
-			timeToNanos(file.Atime), timeToNanos(file.Mtime),
-			timeToNanos(file.Ctime), timeToNanos(file.CreationTime),
+			sqlcodec.TimeToFiletime(file.Atime), sqlcodec.TimeToFiletime(file.Mtime),
+			sqlcodec.TimeToFiletime(file.Ctime), sqlcodec.TimeToFiletime(file.CreationTime),
 			payloadIDPtr, linkTargetPtr, deviceMajor, deviceMinor,
 			file.Hidden, aclJSON, easJSON, objectIDArg,
 			deletedAtArg, file.OriginalPath, file.DeletedBy,
@@ -412,8 +413,8 @@ func (tx *sqliteTransaction) PutFile(ctx context.Context, file *metadata.File) e
 		if _, err := tx.tx.Exec(ctx, insertQuery,
 			file.ID, file.ShareName,
 			file.Type, file.Mode, file.UID, file.GID, file.Size,
-			timeToNanos(file.Atime), timeToNanos(file.Mtime),
-			timeToNanos(file.Ctime), timeToNanos(file.CreationTime),
+			sqlcodec.TimeToFiletime(file.Atime), sqlcodec.TimeToFiletime(file.Mtime),
+			sqlcodec.TimeToFiletime(file.Ctime), sqlcodec.TimeToFiletime(file.CreationTime),
 			payloadIDPtr, linkTargetPtr, deviceMajor, deviceMinor,
 			file.Hidden, aclJSON, easJSON, objectIDArg,
 			deletedAtArg, file.OriginalPath, file.DeletedBy,
@@ -688,10 +689,10 @@ func (tx *sqliteTransaction) ListChildren(ctx context.Context, dirHandle metadat
 			UID:          uint32(uid),
 			GID:          uint32(gid),
 			Size:         uint64(size),
-			Atime:        nanosToTime(atime),
-			Mtime:        nanosToTime(mtime),
-			Ctime:        nanosToTime(ctime),
-			CreationTime: nanosToTime(creationTime),
+			Atime:        sqlcodec.FiletimeToTime(atime),
+			Mtime:        sqlcodec.FiletimeToTime(mtime),
+			Ctime:        sqlcodec.FiletimeToTime(ctime),
+			CreationTime: sqlcodec.FiletimeToTime(creationTime),
 			Hidden:       hidden,
 		}
 		if len(objectIDRaw) > 0 {
@@ -707,15 +708,15 @@ func (tx *sqliteTransaction) ListChildren(ctx context.Context, dirHandle metadat
 		// Recycle-bin metadata (#190): carried on DirEntry.Attr so trash
 		// enumeration via listing reflects recycle state without a re-read,
 		// matching the pool-query path. deleted_at is BIGINT unix-nanoseconds;
-		// decode via nanosToTime.
+		// decode via sqlcodec.FiletimeToTime.
 		if deletedAt.Valid {
-			t := nanosToTime(deletedAt.Int64)
+			t := sqlcodec.FiletimeToTime(deletedAt.Int64)
 			attr.DeletedAt = &t
 		}
 		attr.OriginalPath = originalPath
 		attr.DeletedBy = deletedBy
 
-		// Refs #532 (PR #536 review): mirror fileRowToFileWithNlink — soft
+		// Refs #532 (PR #536 review): mirror sqlcodec.FileRowToFileWithNlink — soft
 		// failure on malformed ACL JSON, same as the pool-query path.
 		if len(aclJSON) > 0 {
 			var fileACL acl.ACL
@@ -1161,10 +1162,10 @@ func (tx *sqliteTransaction) CreateRootDirectory(ctx context.Context, shareName 
 				UID:          uint32(existingUID),
 				GID:          uint32(existingGID),
 				Size:         uint64(size),
-				Atime:        nanosToTime(atime),
-				Mtime:        nanosToTime(mtime),
-				Ctime:        nanosToTime(ctime),
-				CreationTime: nanosToTime(creationTime),
+				Atime:        sqlcodec.FiletimeToTime(atime),
+				Mtime:        sqlcodec.FiletimeToTime(mtime),
+				Ctime:        sqlcodec.FiletimeToTime(ctime),
+				CreationTime: sqlcodec.FiletimeToTime(creationTime),
 				Hidden:       hidden,
 			},
 		}, nil
@@ -1201,10 +1202,10 @@ func (tx *sqliteTransaction) CreateRootDirectory(ctx context.Context, shareName 
 		int32(uid),                        // uid
 		int32(gid),                        // gid
 		int64(0),                          // size
-		timeToNanos(now),                  // atime
-		timeToNanos(now),                  // mtime
-		timeToNanos(now),                  // ctime
-		timeToNanos(now),                  // creation_time
+		sqlcodec.TimeToFiletime(now),      // atime
+		sqlcodec.TimeToFiletime(now),      // mtime
+		sqlcodec.TimeToFiletime(now),      // ctime
+		sqlcodec.TimeToFiletime(now),      // creation_time
 		nil,                               // content_id (NULL for directories)
 		nil,                               // link_target (NULL)
 		nil,                               // device_major (NULL)
@@ -1372,7 +1373,7 @@ func (tx *sqliteTransaction) GetFileByPayloadID(ctx context.Context, payloadID m
 	`
 
 	row := tx.tx.QueryRow(ctx, query, string(payloadID))
-	file, err := fileRowToFileWithNlink(row)
+	file, err := sqlcodec.FileRowToFileWithNlink(row)
 	if err != nil {
 		return nil, mapDBError(err, "GetFileByPayloadID", string(payloadID))
 	}
